+++ aspect-templates/dap-resource.js
@@ -1,1 +1,3 @@
-return distribution;
+// remove collection field (a reference to collection / dataset level info)
+const { collection, ...restFields } = distribution;
+return restFields;