+++ aspect-templates/dcat-distribution-strings.js
@@ -1,9 +1,12 @@
-var moment = libraries.moment;
+const moment = libraries.moment;
+// distribution structure see `FileWithExtraFields` in src/DapTypes
 return {
     title: distribution.name || distribution.id,
     license: distribution.licence || undefined,
     accessURL: distribution.accessURL || undefined,
-    downloadURL: distribution.downloadURL || undefined,
+    // DAP doesn't allow direct ad-hoc download anymore. 
+    // We will hide the download URL for now
+    // downloadURL: distribution.downloadURL || undefined,
     mediaType: distribution.mediaType || undefined,
     format: distribution.format || undefined,
     description: distribution.description || undefined,